requirements.txt
PyPDF2
python-docx
easyocr
transformers
torch 
librosa
pydub